move common fixtures to common module
