+++ nhais-adaptor/adaptor/outgoing/common/tests/fixtures.py
